# backend/api/swagger.yaml
openapi: 3.0.0
info:
  title: Kelo BNPL Platform API
  description: API for the Kelo Buy Now, Pay Later platform - A Web3-powered BNPL solution for the Kenyan market
  version: 1.0.0
  contact:
    name: Kelo Team
    email: api@kelo.ke
    url: https://kelo.ke
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.kelo.ke/v1
    description: Production server
  - url: https://staging-api.kelo.ke/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server
paths:
...
  /stores:
    get:
      summary: List all merchant stores
      description: Get a list of all merchant stores available on the platform.
      tags:
        - Merchant Stores
      responses:
        '200':
          description: Stores retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MerchantStore'
    post:
      summary: Create a new merchant store
      description: Allows a merchant to create a new store.
      tags:
        - Merchant Stores
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MerchantStore'
      responses:
        '201':
          description: Store created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantStore'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /stores/{id}:
    get:
      summary: Get merchant store details
      description: Get detailed information about a specific merchant store.
      tags:
        - Merchant Stores
      parameters:
        - name: id
          in: path
          required: true
          description: Store ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Store details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantStore'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      summary: Update a merchant store
      description: Allows a merchant to update their store's details.
      tags:
        - Merchant Stores
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Store ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MerchantStore'
      responses:
        '200':
          description: Store updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantStore'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '400':
          $ref: '#/components/responses/BadRequestError'
...
  /orders:
    get:
      summary: List user orders
      description: Get a list of orders for the current user.
      tags:
        - Orders
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: Create a new order
      description: Create a new order from a list of products.
      tags:
        - Orders
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /orders/{id}:
    get:
      summary: Get order details
      description: Get detailed information about a specific order.
      tags:
        - Orders
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # Liquidity Pools
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MerchantStore:
      type: object
      properties:
        id:
          type: string
          format: uuid
        merchant_id:
          type: string
          format: uuid
        store_name:
          type: string
        store_url:
          type: string
          format: uri
        store_address:
          type: string
        logo_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        merchant_store_id:
          type: string
          format: uuid
        total_amount:
          type: number
          format: float
        status:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderItem:
      type: object
      properties:
        order_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
        price_at_purchase:
          type: number
          format: float
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Bad request
        message:
          type: string
          example: Invalid request parameters
        code:
          type: integer
          example: 400
        details:
          type: object
          description: Additional error details
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...